requirements.txt
numpy==1.17.0
pandas==0.23.4
scikit-learn==0.21.3
tensorflow-gpu==1.13.1
